feat: full task crud

+++ api/utils/user_utils.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm import Session
 from database.models.user_model import User
-from schemas.user_schema import UserCreate
+from schemas.user_schema import UserCreate, UserUpdate
+from fastapi import HTTPException
 
 
 def get_user(db: Session, user_id: int):
@@ -16,26 +17,26 @@ def get_users(db: Session, skip: int = 0, limit: int = 100):
 
 
 def create_user(db: Session, user: UserCreate):
-    db_user = User(username=user.username, status=user.status, password=user.password)
+    db_user = User(username=user.username, profile=user.profile, password=user.password)
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
     return db_user
 
 
-def update_user(db: Session, user_id: User, user: UserCreate):
-    db_user = db.query(User).filter(User.id == user_id).first()
-    if user.username:
-        db_user.username = user.username
-    if user.password:
-        db_user.password = user.password
-    if user.status:
-        db_user.status = user.status
+def update_user(db: Session, user_id: int, user_data: dict):
+    db_user = get_user(db=db, user_id=user_id)
+    if db_user is None:
+        raise HTTPException(status_code=404, detail="User not found")
+    for field, value in user_data.items():
+        setattr(db_user, field, value)
     db.commit()
     return db_user
 
 
-def delete_user(db: Session, user_id: User):
-    db_user = db.query(User).filter(User.id == user_id).first()
+def delete_user(db: Session, user_id: int):
+    db_user = get_user(db=db, user_id=user_id)
+    if db_user is None:
+        raise HTTPException(status_code=404, detail="User not found")
     db.delete(db_user)
     db.commit()
